requirements.txt
discord.py
python-dotenv
qwen-tts
torch
soundfile